Test boudanry cost

+++ BillionScaleSearch/Optkmeans/Optkmeans.h
@@ -13,8 +13,8 @@
 #include <omp.h>
 
 // Need to set this correctly when running million/billion scale datasets
-typedef float DataType;
-//typedef uint8_t DataType;
+//typedef float DataType;
+typedef uint8_t DataType;
